Changed OpCode ordinal to byte.

+++ src/main/java/com/karalabe/iris/OpCode.java
@@ -5,27 +5,27 @@ package com.karalabe.iris;
  */
 /*@formatter:off*/
 enum OpCode {
-    INIT(0),           /** Connection initialization */
-    BROADCAST(1),      /** Application broadcast */
-    REQUEST(2),        /** Application request */
-    REPLY(3),          /** Application reply */
-    SUBSCRIBE(4),      /** Topic subscription */
-    PUBLISH(5),        /** Topic publish */
-    UNSUBSCRIBE(6),    /** Topic subscription removal */
-    CLOSE(7),          /** Connection closing */
-    TUNNEL_REQUEST(8), /** Tunnel building request */
-    TUNNEL_REPLY(9),   /** Tunnel building reply */
-    TUNNEL_DATA(10),   /** Tunnel data transfer */
-    TUNNEL_ACK(11),    /** Tunnel data acknowledgment */
-    TUNNEL_CLOSE(12);  /** Tunnel closing */
+    INIT((byte)0),           /** Connection initialization */
+    BROADCAST((byte)1),      /** Application broadcast */
+    REQUEST((byte)2),        /** Application request */
+    REPLY((byte)3),          /** Application reply */
+    SUBSCRIBE((byte)4),      /** Topic subscription */
+    PUBLISH((byte)5),        /** Topic publish */
+    UNSUBSCRIBE((byte)6),    /** Topic subscription removal */
+    CLOSE((byte)7),          /** Connection closing */
+    TUNNEL_REQUEST((byte)8), /** Tunnel building request */
+    TUNNEL_REPLY((byte)9),   /** Tunnel building reply */
+    TUNNEL_DATA((byte)10),   /** Tunnel data transfer */
+    TUNNEL_ACK((byte)11),    /** Tunnel data acknowledgment */
+    TUNNEL_CLOSE((byte)12);  /** Tunnel closing */
 
-    private final int ordinal;
+    private final byte ordinal;
 
-    private OpCode(int ordinal) {
+    private OpCode(byte ordinal) {
         this.ordinal = ordinal;
     }
 
-    public int getOrdinal() {
+    public byte getOrdinal() {
         return ordinal;
     }
 }
